sparta_news/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
billiard==4.2.0
blis==0.7.11
catalogue==2.0.10
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
decorator==5.1.1
distro==1.9.0
Django==4.2.11
django-crontab==0.7.1
django-extensions==3.2.3
django-redis==5.4.0
django-seed==0.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.1
executing==2.0.1
Faker==24.11.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
kombu==5.3.7
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.1.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
murmurhash==1.0.10
numpy==1.26.4
openai==1.23.6
packaging==24.0
parso==0.8.4
pexpect==4.9.0
preshed==3.0.9
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
redis==5.0.4
requests==2.31.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.0
srsly==2.4.8
stack-data==0.6.3
thinc==8.2.3
toposort==1.10
tqdm==4.66.2
traitlets==5.14.3
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
